Fixed grotto generator set seed

+++ Source/Core/Gen5/States/HiddenGrottoState.hpp
@@ -28,7 +28,7 @@ class HiddenGrottoState
 public:
     HiddenGrottoState() = default;
 
-    HiddenGrottoState(u64 seed, u32 advances, u8 group, u8 slot, u8 gender) :
+    HiddenGrottoState(u32 seed, u32 advances, u8 group, u8 slot, u8 gender) :
         seed(seed), advances(advances), group(group), slot(slot), gender(gender)
     {
     }
@@ -59,7 +59,7 @@ public:
     }
 
 private:
-    u64 seed;
+    u32 seed;
     u32 advances;
     u8 group;
     u8 slot;
